move overlay test to own class
