+++ src/main/java/com/cleanroommc/modularui/ClientProxy.java
@@ -7,6 +7,7 @@ import com.cleanroommc.modularui.keybind.KeyBindHandler;
 import com.cleanroommc.modularui.overlay.OverlayManager;
 import com.cleanroommc.modularui.screen.ClientScreenHandler;
 import com.cleanroommc.modularui.test.EventHandler;
+import com.cleanroommc.modularui.test.OverlayTest;
 import com.cleanroommc.modularui.theme.ThemeManager;
 import com.cleanroommc.modularui.theme.ThemeReloadCommand;
 
@@ -34,11 +35,11 @@ public class ClientProxy extends CommonProxy {
         MinecraftForge.EVENT_BUS.register(ClientEventHandler.class);
         MinecraftForge.EVENT_BUS.register(ClientScreenHandler.class);
         MinecraftForge.EVENT_BUS.register(OverlayManager.class);
-        MinecraftForge.EVENT_BUS.register(GuiStack.INSTANCE);
         MinecraftForge.EVENT_BUS.register(KeyBindHandler.class);
 
         if (ModularUIConfig.enabledTestGuis) {
             MinecraftForge.EVENT_BUS.register(EventHandler.class);
+            OverlayTest.init();
         }
 
         DrawableSerialization.init();
